Rename azure{,rm}_security_center_polic* resources
And replace a `select` with `reject` because Paul didn't run Rubocop.
