+++ docs/resources/azurerm_security_center_policies.md.erb
@@ -1,11 +1,11 @@
 ---
-title: About the azure_security_center_policies Resource
+title: About the azurerm_security_center_policies Resource
 platform: azure
 ---
 
-# azure\_security\_center\_policies
+# azurerm\_security\_center\_policies
 
-Use the `azure_security_center_policies` InSpec audit resource to test
+Use the `azurerm_security_center_policies` InSpec audit resource to test
 properties of some or all Azure Security Center Policies.
 
 Security Center Policies are defined for each Resource Group. A Security Center
@@ -15,11 +15,11 @@ Policy called `default` also exists for every subscription.
 
 ## Syntax
 
-An `azure_security_center_policies` resource block uses an optional filter to
+An `azurerm_security_center_policies` resource block uses an optional filter to
 select a group of Security Center Policies and confirm that the expected groups
 exist.
 
-    describe azure_security_center_policies do
+    describe azurerm_security_center_policies do
       ...
     end
 
@@ -29,13 +29,13 @@ exist.
 
 ### Check for a Security Center Policy
 
-    describe azure_security_center_policies do
+    describe azurerm_security_center_policies do
       its('names') { should include 'default' }
     end
 
 ### Assert default Security Center Policy exists
 
-    describe azure_security_center_policies.where('name' => 'default')
+    describe azurerm_security_center_policies.where('name' => 'default')
       it { should exist }
     end
 
@@ -51,7 +51,7 @@ Filters the results to include only those Security Center Policies that match
 the given name. This is a string value.
 
     # default should always exist
-    describe azure_security_center_policies.where('name' => 'default')
+    describe azurerm_security_center_policies.where('name' => 'default')
       it { should exist }
     end
 
@@ -67,12 +67,12 @@ The control will pass if the filter returns at least one result. Use
 `should_not` if you expect zero matches.
 
     # default should always exist
-    describe azure_security_center_policies.where('name' => 'default')
+    describe azurerm_security_center_policies.where('name' => 'default')
       it { should exist }
     end
 
     # this security center policy should not exist
-    describe azure_security_center_policies.where('name' => 'DoesNotExist')
+    describe azurerm_security_center_policies.where('name' => 'DoesNotExist')
       it { should_not exist }
     end
