+++ libraries/azure_security_center_policies.rb
@@ -1,28 +1,20 @@
 # frozen_string_literal: true
 
-require 'azurerm_resource'
+require 'azurerm_security_center_policies'
 
-class AzureSecurityCenterPolicies < AzurermResource
+class AzureSecurityCenterPolicies < AzurermSecurityCenterPolicies
   name 'azure_security_center_policies'
-  desc 'Verifies settings for Security Center'
+  desc '[DEPRECATED] Please use the azurerm_security_center_policies resource'
   example <<-EXAMPLE
     describe azure_security_center_policies do
       its('policy_names') { should include('default') }
     end
   EXAMPLE
 
-  FilterTable.create
-             .add_accessor(:entries)
-             .add_accessor(:where)
-             .add(:exists?) { |obj| !obj.entries.empty? }
-             .add(:policy_names, field: 'name')
-             .connect(self, :table)
-
-  def to_s
-    'Security Policies'
-  end
-
-  def table
-    @table ||= client.security_center_policies
+  def initialize
+    warn '[DEPRECATION] The `azure_security_center_policies` resource is ' \
+         'deprecated and will be removed in version 2.0. Use the ' \
+         '`azurerm_security_center_policies` resource instead.'
+   super
   end
 end
